* Added logs and traces tablename utiliy * Refactoring changes
