+++ exporter/azuredataexplorerexporter/factory.go
@@ -113,7 +113,7 @@ func createTracesExporter(
 	}
 	// call the common exporter function in baseexporter. This ensures that the client and the ingest
 	// are initialized and the metrics struct are available for operations
-	amp, err := newExporter(adxCfg, set.Logger, LogsType)
+	amp, err := newExporter(adxCfg, set.Logger, TracesType)
 
 	if err != nil {
 		return nil, err
